!!! NOT TESTED !!!
*** AIP ***
-> adding the aipid when executing the aip conditions and actions, will be easier for logs and patching,


*** Gem Quest // gm command ***
the gem quest begins at 19:56:00 to 20:01:00 and ends at 22:00:00.
Monsters (junon's xxx) are spawned when you kill regular monsters after 20:00:00, see export.log (made by massexporter option) to see
which regular monster can spawn those when killed.
When one of those monster is spawned, a counter is added in the NPC 1104.

The gm command is more like a debug thing than a real alternative since the quest won't close on itself once manually activated.
I do this because it's very easy to change the condtions in NPC_1104.AIP file, idle section, you can do that server side only if you like.
-> gm command made /gemquest option [value1] [value2]
--> start nbminutes autoreset:
---> will start the gem quest in now + nbminutes, and can auto reset the number of monster spawned (autoresetat can be 0 or more, 0 means desactivated),
---> (/gemquest start 5 10), WARNING, if you do this, the quest WON'T end on hiw own, you'll have to close it yourself using "stop" option later,
---> autoresetat is used because in the quest after 60 (not sure about the amount) "Junon's" monsters are spawned, the quest stops. So if you set autoresetat at 10 for example,
---> the monster spawn counter will be set back to 0 after 10 have been spawned.
--> stop:
---> will stop the gem quest if started manually (/gemquest stop),
--> info:
---> will give some info on the quest (started or not, how many monsters killed since there is a maximum...), "/gemquest info",
--> reset:
---> will reset the spawn counter (see start option for explanation),

So to make it short, if you want to start the gemquest in 5 minutes and that you want it to last until you stop it, set autoresetat to 20 or something so you'll be sure the "Junon's" monsters will continue to spawn:
/gemquest start 5 20
If you want to start the gemquest in 5 minutes and you don't want the Junon's monsters to continue to spawn after the amount they're supposed to:
/gemquest start 5 0
If the gem quest is active and you want to "reset" the spawn counter:
/gemquest reset

Note:
/gemquest info
will tell you the actual spawn counter and if the autoresetat is set and its value.